backend/api/urls.py
from django.urls import path, include
from rest_framework import routers

from api.views import TagsViewSet, RecipesViewSet, \
    IngredientsViewSet, CustomUserViewSet


router = routers.DefaultRouter()
router.register('users', CustomUserViewSet, basename='users')
router.register('tags', TagsViewSet, basename='tags')
router.register('recipes', RecipesViewSet, basename='recipes')
router.register('ingredients', IngredientsViewSet, basename='ingredients')


urlpatterns = [
    path('auth/', include('djoser.urls.authtoken')),
    path('', include(router.urls)),
]
